+++ MangaMark_Vue/src/popup/Popup.vue
@@ -1,27 +1,20 @@
 <script setup lang="js">
-import { ref, watch, onMounted, onUnmounted } from 'vue'
-import logWithTimestamp from '../js/utils/logWithTimestamp'
+import { ref, onMounted } from 'vue'
 import getCurrentTab from '../js/utils/getCurrentTab'
 import { packageName, version } from '../js/global'
-import '../assets/css/popup.css'
+import '../assets/css/bulma.min.css'
+import Options from '../options/Options.vue'
 
 const id = ref('')
 const api = ref('')
 const currentTab = ref(null)
 const gearSrc = ref('')
+const switchOption = ref(false)
 
 document.title = packageName + ' - Popup'
 
 function openOption() {
-  chrome.storage.sync.set({ TAB_ID: currentTab.value.id })
-  chrome.runtime
-    .openOptionsPage()
-    .then(() => {
-      logWithTimestamp('Open options page')
-    })
-    .catch((err) => {
-      logWithTimestamp(err)
-    })
+  switchOption.value = !switchOption.value
 }
 
 function saveLogin() {
@@ -39,25 +32,53 @@ onMounted(async () => {
   chrome.storage.sync.get(['ID', 'API'], (result) => {
     id.value = result.ID || ''
     api.value = result.API || ''
+    switchOption.value = id.value === '' || api.value === ''
   })
+
   currentTab.value = await getCurrentTab()
   gearSrc.value = chrome.runtime.getURL('icons/gear.png')
 })
 </script>
 
 <template>
-  <main>
-    <h1>{{ packageName }} v{{ version }}</h1>
-    <span id="option" @click="openOption"><img width="26" :src="gearSrc" /></span>
-    <div id="login">
-      <label for="id">ID</label>
-      <input type="text" id="id" name="id" placeholder="Your ID" v-model="id" />
+  <main class="has-background-light py-3 px-3" style="width: 380px">
+    <div class="container">
+      <div style="display: flex; justify-content: space-between">
+        <h2 class="subtitle">{{ packageName }} v{{ version }}</h2>
+        <span @click="openOption" style="margin-left: auto; cursor: pointer"
+          ><img width="26" :src="gearSrc"
+        /></span>
+      </div>
+      <template v-if="switchOption">
+        <div class="field">
+          <label class="label">ID</label>
+          <div class="control">
+            <input class="input is-rounded" type="text" placeholder="Your ID" v-model="id" />
+          </div>
+        </div>
 
-      <label for="api">API Key</label>
-      <input type="text" id="api" name="api" placeholder="Your API Key" v-model="api" />
-      <input @click="saveLogin" id="save" type="button" value="Save" />
+        <div class="field">
+          <label class="label">API Key</label>
+          <div class="control">
+            <input class="input is-rounded" type="text" placeholder="Your API Key" v-model="api" />
+          </div>
+        </div>
+
+        <div class="field is-grouped is-justify-content-center">
+          <div class="control">
+            <button @click="saveLogin" class="button is-link">Save</button>
+          </div>
+        </div>
+      </template>
+      <template v-else>
+        <Options />
+      </template>
     </div>
   </main>
 </template>
 
-<style></style>
+<style>
+/* #login-form {
+  width: 200px;
+} */
+</style>